requirements.txt
google-cloud-vision==3.4.5
google-cloud-storage==2.13.0
google-cloud-firestore==2.13.1
gspread==5.12.2